+++ assets/javascripts/kitten/components/information/toggletip/stories.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { Toggletip } from './index'
 import { DocsPage } from 'storybook/docs-page'
 import { Button } from 'kitten'
+import NumberFormat from 'react-number-format'
 
 export default {
   title: 'Information/Toggletip',
@@ -20,6 +21,7 @@ const args = {
   modifier: 'info',
   children: 'The text of my Toggletip.',
   targetElement: undefined,
+  bubbleColorText: null,
 }
 
 const argTypes = {
@@ -43,6 +45,10 @@ const argTypes = {
   targetElement: {
     name: 'targetElement',
     control: 'text',
+  },
+  bubbleColorText: {
+    name: 'bubbleColorText',
+    control: 'color',
   },
 }